Skip Exodus output in examples if its unavailable.
Rationale: we presently set a default file name for the Exodus output in
all of the parameter files. Removing that default setting would be a
breaking change since we would, by default, disable all exodus
output. At the same time, not *all* examples require Exodus, so we do
not want to unconditionally require it and break current setups based on
non-Exodus examples. Hence, we compromise: we keep the default file name
but log a warning (and not write any Exodus output) if Exodus output is
requested but not possible.