Initialize IBStandardInitializer in its ctor.
This fixes an issue related to the current failure of the CIB/ex0 test:
if, in that test, output is disabled, then init() is never run, which
leads to failed assertions later on. Since init() does not rely on any
data not set in the constructor (i.e., the class members are set by
getFromInput) we can run it there.