Run one test at stricter tolerances.
Commit e6aeeb564ca changed the way we do a particular L2 projection in
IBFE/explicit/ex0: instead of using a given initial guess for the
solution we now zero the vector. This caused the computed solution to be
slightly different from the known value.

This commit changes the test to run at a much stricter numerical
tolerance, and compares it to values computed (with the same tolerances)
before e6aeeb564ca.