Allow versions of PETSc newer than 3.9.
PETSc 3.10 will be out soon and we are compatible with it.

The PETSc interfaces we use are quite stable and are unlikely to change in the
near future, so there is no reason to specify a maximum version: we should
simply document what works and let users use new versions if they like.