Fix alignment of Eigen objects.
Eigen will automatically use AVX512 and other vectorized instruction
sets if they are available. This causes segmentation faults (on AMD64,
at least) when objects are not aligned to 128-bit (or better)
addresses. This commit gets around this by specifying alignment
requirements (or, for std::vector [1], a custom allocator) for objects
with Eigen members.

[1] http://eigen.tuxfamily.org/dox-devel/group__TopicStlContainers.html