Add FEMapCache.
Similar to FECache. Since libMesh will always recompute the shape values
when it encounters a new cell (even if they have not changed) we have to
decouple the computation of the Jacobians from the computation of the
shape values. Hence, we need a second cache of FEMap objects (which
compute JxW values).