Enable a faster code path in index lookups.
If we explicitly localize the vector through the libMesh interface then
the local sizes are cached: i.e., we no longer need to call
VecGetOwnershipRange repeatedly when looking up local indexes since
libMesh assumes the value is constant.

Since libMesh uses exactly the same localization strategy that we do its
worthwhile to just let it manage this part so that we can use its own
optimizations when looking up indices.