Rearrange most workload estimation code.
This commit massively overhauls the work estimation code. The new
implementation assumes that one parent hierarchy integrator with
multiple children is ultimately responsible for computing work
estimates: each child is asked, in sequence, to compute their own
workload estimates and add them into the specified workload index
variable. This solves the problem where multiple objects might try to
zero the workload estimate array: since the parent integrator owns it,
zeroing the data must be the parents responsibility.

Since the parent passes the correct index as a function argument, the
correct index can be propagated to things like LDataManager and
IBFEMethod, which eliminates the need to extract the index and then
separately supply it to those objects.