Switch to addWorkloadEstimate in non-hierarchies.
Since these classes don't ultimately own the data they should only be
expected to sun into the given variable of the given hierarchy with
their workload estimates.

Unfortunately, since, e.g., IBHierarchyIntegrator calls
IBStrategy::addWorkloadEstimate, this has to be an incompatible change:
the semantics of addWorkloadEstimate are slightly different (it does not
assume that a load balancer is attached or that the workload index is
already stored) and there is no way to determine if a user code
implements addWorkloadEstimate or updateWorkloadEstimates.