Cache interpolated positions.
We spend about 5% of our total run time in the heart model calling this
function. This commit reduces the number of calls per cell from five to
two (i.e., instead of calling it three times per cell in interpWeighted
we call it once and instead of twice in spread now we just call it one
once).