Improve performance of FEDataManager by unrolling.
These percentages all come from running the heart model in callgrind
with one processor. This patch lowers the total number of instructions
by 2.5%.  About 4.1% of total instructions were spent in interpWeighted
itself (or in inlined functions): this patch lowers that to 1.6% with
the new integrate_cell_rhs function taking up 0.7% of total
instructions.