Change the way we set up IB ghosted vectors.
The current strategy, based on FEDataManager::buildIBGhostedVector,
stores a vector inside FEDataManager and a pointer to that vector inside
IBFEMethod. This commit changes the way we do things so that the vectors
are actually owned by IBFEMethod itself.

This change is necessary since some components require ghosted RHS
vectors and others require ghosted solution vectors: it is easiest to
simply keep cloning these things whenever we need them instead of
worrying about how we should interpret the vector owned by
FEDataManager.

While we are here: set up the projected velocity RHS vector with IB
ghosted entries. This changes nothing at the moment but is needed in a
future PR.