Fix link order dependencies with silo and HDF5.
This commit ensures that we always explicitly link with libhdf5,
libhdf5_hl, and silo, and puts the order of libraries in the correct
order (i.e., HDF5 first, then silo, which depends on HDF5).

Unfortunately, this change necessitated changes in how we configure
PETSc, libMesh, and HYPRE to keep compatibility with some more unusual
situations (e.g., libMesh must be configured first since it will
overwrite PETSC_DIR when appropriate).

This is also the first commit in a series which will remove the PACKAGE_
infrastructure from IBAMR: the current state of autotools is good enough
that we should just use standard macros.