Make IBFEMethod::initializeLevelData a no-op.
We initialize FE data in a different way in this class and this
secondary entry point has really confusing semantics, since it can be
called when we regrid both normally and when we (in the next patch)
regrid the scratch hierarchy. Since this doesn't do anything useful
anyway we can just drop it.