Fix restarts on the scratch hierarchy.
A bug in SAMRAI prevents the usage of GriddingAlgorithm objects that do
not correspond to a restart database entry (i.e., one cannot dynamically
create GriddingAlgorithm objects in a restarted simulation). Get around
this by setting up the GriddingAlgorithm object and friends in the
constructor of IBFEMethod so that they are always available.