pseudo code for using the initial configuration
We want to use the initial configuration of the structure for FSI for
structures that shouldn't move (e.g., flow past a cylinder). Try this
out by always using the initial configuration (stored as
"initial_configuration" in the X_system) in both velocity interpolation
and force spreading.

This code doesn't work but shows what we need to do.