Fix IB-5.
Like the other wider IB stencils (IB-4 and up) IB-5 inlines the
computation of the kernel and computes the full stencil regardless of
where we are in the box. However, it indexes the computed stencil from 0
regardless of whether or not we actually use the full stencil - this
results in using the wrong coefficients when we don't use the full
stencil.