Change the default settings for libMesh remeshing.
The old version was based on an incorrect assumption that the
nonlocality of libMesh data was the cause of performance problems - in
fact, the problem was/is the number of IB points per processor. We
should not repartition the Lagrangian data unless the user explicitly
asks for it since this will actually degrade performance in the
calculation of forces on the Lagrangian mesh.