Fix SAMRAIGhostDataAccumulator + SideData.
The box returned by getGhostBox is for the patch: it is not the offset box
required to correctly use side data. Using the wrong box caused the function to
skip the last column which results in wrong accumulation values.

In the modified test output, note that 6 really should be the maximum value
since that entry shows up four times on the patch interior (twice on each patch
due to periodicity) and twice more on one patch in ghost entries.