Fix a busy wait problem in attest.
This prevents us from running python at 100% when we have more available
processors than work. This can happen when, for example, we run

    ./attest -j6 -R 'mpirun=4'

here we will never fully utilize the number of allocated cores, so the loop will
endlessly attempt to schedule more jobs without waiting on the lone worker to
finish first.