Use a different algorithm for nodal quadrature interpolation.
We previously de facto interpolated the hanging node constraints in a manner
similar to that described in Kormann's 2016 paper on mass lumping for the
Schrodinger equation. This algorithm is problematic for us, in particular, since
it results in ignoring the values interpolated at hanging nodes while spreading
forces from hanging nodes. Get around this by properly forming the lumped mass
matrix (which, if there are hanging nodes, is no longer diagonal).

We should write a clearer discussion than this for the upcoming mass lumping
paper, but the reason this newer approach is preferrable is that we are now
using the same mass matrix C^T M C for both spreading and interpolation. For
interpolation we use the new lumped (and not necessarily diagonal) mass matrix
to solve the linear system. For spreading we use the unconstrained diagonal mass
matrix, but we only use it with finite element solution vectors (i.e., vectors x
where C x = x): hence the inner product induced by C^T M C is equal to that of
just M in the nodal quadrature case. Similarly, we use the same JxW values and
point locations for both spreading and interpolation (they come from the
diagonal mass matrix).