Get rid of an extra warning.
This does not work as intended since if absolutely nothing is
specified (like in most of the tests) then timer_dump_interval_key_name
is an empty string - lets just avoid the problem by getting rid of
it. The problem is obvious, anyway - if no timer information is
specified the TimerManager will print things we don't want.

Furthermore - we get problems in some tests if we use NullDatabase in
debug mode, but we can just supply a null pointer instead.