Fix the merging load balancer.
The extra insert statement doesn't make any sense: its a set so it
already contains its first element, and if it is an empty set then this
creates an invalid object. In particular, since some processors can end
up with zero boxes in very strange situations, this causes nonsense
boxes to be created in the output as a result.