Transfer data inside the SecondaryHierarchy.
I want to put this in now so that we can go back and add more timers later -
that will make it easier to justify the choice of multiple Eulerian data
partitionings in the load balancing paper.

In particular, I want to merge this now since this is an API change and we don't
have to worry about compatibility with this class yet since it is new in 0.9.