Update bundled boost version to 1.76.
I have tried to use a smaller subset of boost to keep downloads reasonably small
- after all, we only use four headers from boost (multi_array, roots, round, and
demangle).

To compute the minimal subset, I copied and pasted the four headers we use and
then ran the following shell script:

    #!/bin/bash
    set -e

    for i in $(seq 0 1000)
    do
        echo "round ${i}"
        for FILE in $(bash ./test-boost-2.sh)
        do
            mkdir -p $(dirname "boost/${FILE}")
            echo "file = ${FILE}"
            cp temp/boost_1_76_0/${FILE} ./boost/${FILE}
        done
    done

which depends on this shell script:

    #!/bin/bash
    # set -e

    for FILE in $(find ./boost -type f)
    do
        /usr/local/bin/g++-11 -DBOOST_ALL_NO_LIB -I$(pwd)/boost -E ${FILE} -o test.o 2>&1 | grep 'No such file' | awk -F: '{print $5}'
    done | grep boost | sort -u | grep -v 'MPL_PREPROCESSED_HEADER'

to ensure that we can at least preprocess (i.e., all dependent headers are
present) each required header. If we want to add more headers in the future then
we should rerun the first script to add its minimal dependencies. The process
took about 400 'iterations' to converge.