Remove optional communicator arguments.
This came up on the mailing list and, unfortunately, the permanent fix
is quite long.

The issue we have is that we made the MPI communicator argument
implicit. Some versions of MPI have

    typedef unsigned int MPI_Comm;

which means that, with

    template <typename T>
    static T sumReduction(T, IBTK_MPI::comm commiunicator = getCommunicator());
    template <typename T>
    static void sumReduction(T* x, const int n = 1, IBTK_MPI::comm communicator = getCommunicator());

if we do

    unsigned int size = 100;
    double *values = array.begin();
    sumReduction(values, size);

we end up calling the first function instead of the second one (due to
C++ overload resolution rules its a better fit), which is wrong.