Disable checking MPI library linking if we use mpicc etc.
If we use the MPI compiler wrappers then MPI_C_LIBRARIES is empty, as there are
no additional libraries for us to link against. In this case this check will
unconditionally fail since whatever MPI library an external dependency links
against will obviously not be in that empty list.

Hence, only run the check if MPI_C_LIBRARIES has something in it. Regardless of
the link check, we use PETSc's preprocessor logic to verify that we have the
same MPI library version, so we still have pretty good protection against mixing
MPI implementations by accident.