Fix the 'point outside' check with some other examples.
The line

    hierarchy->getPatchLevel(vertex_level_number);

fails inside SAMRAI with a difficult to understand type casting error for some
tests in which the grid is not fully set up at the point when this is called.
Fortunately the fix is pretty easy - we don't actually need to account for the
periodic shift to get the actual position of a vertex, we just need to know
which directions are periodic since those can be ignored when we check for
points outside the physical domain.