INS*HierarchyIntegrator: only compute vorticity when we need it.
This fixes a bug caused by not storing the vorticity in the restart database. In
that case, the solver needed to regrid at the first post-restart timestep and
the vorticity was not available.

This is more efficient than the current implementation since the INS solvers no
longer compute the vorticity at the end of every timestep. However, it does
require (since tagging occurs one level at a time) that we compute the vorticity
once per tagged level, which makes applyGradientDetector() less efficient. We
still come out ahead as long as there are fewer levels than steps between
regrids (which is the usual situation - i.e., even with 4-5 levels, it is very
rare that we regrid more freqently than every 4-5 timesteps).