IBHierarchyIntegrator etc.: refactor sources derived from IBStrategy.
This is the first of two commits which will fix a bug where we didn't
correctly update sources in calls to
INSHierarchyIntegrator::regridProjection(). To begin, this commit
makes it possible to compute sources outside of the normal timestepping
loop and also modifies IBEulerianSourceFunction to work on the entire
PatchHierarchy (instead of single patches) at once.