attest: fix some parsing bugs.
We previously assumed that the base filename (i.e., everything before
the first '.') did not contain the strings 'expect_error', 'mpirun' , or
'restart' - this is a problem with the restart cleaner tests. In all
cases, these are special commands which cannot appear in the basename,
so we can fix it by just ignoring the basename.